+++ src/main/java/kg/bishkek/fucent/fusent/dto/ProductUpdateDtos.java
@@ -2,12 +2,13 @@ package kg.bishkek.fucent.fusent.dto;
 
 // kg.bishkek.fucent.fusent.dto.ProductUpdateDtos.java
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class ProductUpdateDtos {
     public record VariantPriceStockUpdate(
             @NotNull UUID variantId,
-            Double price, Integer stockQty
+            BigDecimal price, Integer stockQty
     ) {}
     public record ProductActivateRequest(@NotNull UUID productId, boolean active) {}
 }
